// TaskManagementWebClient/src/models/task-detail.model.ts
export class TaskDetailViewModel {
    TaskDetailId: number = 0;
    Comment: string = null;
    CommentDateTime: Date = null;
    IsReplyed: boolean = false;
    AllTaskId: number = 0;
    EmployeeId: number = 0;
}

export class TaskDetail {
    taskDetailId: number = 0;
    comment: string = null;
    commentDateTime: Date = null;
    isReplyed: boolean = false;
    allTaskId: number = 0;
    employeeId: number = 0;
}